Aprendendo a usar Lombok

+++ Tests in Progress/SpringBoot/drogaria/src/main/java/br/com/testes/drogaria/domain/Produto.java
@@ -13,7 +13,7 @@ public class Produto {
 	
 	private String nome;
 	
-	private Short quantidade;
+	private Byte quantidade;
 	
 	private BigDecimal preco;
